[tool/package] Rename nickname -> displayname
Both packages and tools have the two attributes nickname and name. This
was confusing. I renamed the nickname into displayname to make it
clearer that this it the name that shows up in the different templates.

Even better I think would be to later refactor so that the `name`
attribute is used in the templates and then use something else for the
internal other name (once I figured out what it is for).