Make sure core.fasta is on the worknode.
This is an ugly workaround from the decision to remove the fido_tools
part. The correct solution is to break out more into an extra package,
there is a parse_mdrscan.py that before wasn't in the fido_tools, there
are some ugly code sharing here that need to be formalized.