[mdr] Fix template search.
The mdr view for specific families need to know the absolute path to the
template. Using the Django template loader didn't work for this anymore
so we have to hardwire the path to the template file.

This is a misfeature and we need to work on removing the special
handling of this. Either by introducing paging or by dynamically loading
content with javascript. Preferably it should page if the user has no
javascript and dynamically fetch the extra information if javascript is
available.